refactor(base-payment-lib): Cleanup util files
EC-106
