+++ gradle/libs.versions.toml
@@ -3,7 +3,7 @@ android-compileSdk = "34"
 android-minSdk = "21"
 android-targetSdk = "34"
 kotlin = "1.8.20"
-android-gradle-plugin = "8.4.2"
+android-gradle-plugin = "8.2.2"
 dokka = "1.7.20"
 androidx-lifecycle = "2.7.0"
 androidx-camera = "1.3.1"
