+++ internal-payment-sdk/sdk/src/main/AndroidManifest.xml
@@ -2,7 +2,7 @@
 <manifest xmlns:android="http://schemas.android.com/apk/res/android">
     <application android:label="@string/app_name">
         <provider
-            android:name="net.gini.android.internal.payment.utils.PaymentFileProvider"
+            android:name="net.gini.android.internal.payment.util.PaymentFileProvider"
             android:authorities="${applicationId}.internal.payment.fileprovider"
             android:exported="false"
             android:grantUriPermissions="true">
@@ -10,7 +10,7 @@
                 android:name="android.support.FILE_PROVIDER_PATHS"
                 android:resource="@xml/file_paths" />
         </provider>
-        <receiver android:name="net.gini.android.internal.payment.utils.ShareWithBroadcastReceiver" android:exported="false"/>
+        <receiver android:name="net.gini.android.internal.payment.util.ShareWithBroadcastReceiver" android:exported="false"/>
 
     </application>
