+++ internal-payment-sdk/sdk/src/main/java/net/gini/android/internal/payment/review/installApp/InstallAppViewModel.kt
@@ -9,7 +9,7 @@ import kotlinx.coroutines.launch
 import net.gini.android.internal.payment.paymentComponent.PaymentComponent
 import net.gini.android.internal.payment.paymentComponent.SelectedPaymentProviderAppState
 import net.gini.android.internal.payment.paymentprovider.PaymentProviderApp
-import net.gini.android.internal.payment.utils.BackListener
+import net.gini.android.internal.payment.util.BackListener
 import org.slf4j.LoggerFactory
 
 internal class InstallAppViewModel(private val paymentComponent: PaymentComponent?, val backListener: BackListener?) : ViewModel() {
