+++ internal-payment-sdk/sdk/src/main/java/net/gini/android/internal/payment/review/openWith/OpenWithViewModel.kt
@@ -3,7 +3,7 @@ package net.gini.android.internal.payment.review.openWith
 import androidx.lifecycle.ViewModel
 import androidx.lifecycle.ViewModelProvider
 import net.gini.android.internal.payment.paymentprovider.PaymentProviderApp
-import net.gini.android.internal.payment.utils.BackListener
+import net.gini.android.internal.payment.util.BackListener
 
 internal class OpenWithViewModel private constructor(val paymentProviderApp: PaymentProviderApp?, val backListener: BackListener?): ViewModel() {
     class Factory(private val paymentProviderApp: PaymentProviderApp?, private val backListener: BackListener?) : ViewModelProvider.Factory {
