+++ internal-payment-sdk/sdk/src/main/java/net/gini/android/internal/payment/review/reviewComponent/ReviewView.kt
@@ -5,7 +5,6 @@ import android.text.SpannableStringBuilder
 import android.text.Spanned
 import android.text.style.ImageSpan
 import android.util.AttributeSet
-import android.util.Log
 import android.view.View
 import androidx.constraintlayout.widget.ConstraintLayout
 import androidx.core.graphics.drawable.RoundedBitmapDrawableFactory
@@ -23,15 +22,15 @@ import net.gini.android.internal.payment.paymentComponent.SelectedPaymentProvide
 import net.gini.android.internal.payment.paymentprovider.PaymentProviderApp
 import net.gini.android.internal.payment.review.PaymentField
 import net.gini.android.internal.payment.review.ValidationMessage
-import net.gini.android.internal.payment.utils.amountWatcher
-import net.gini.android.internal.payment.utils.extensions.clearErrorMessage
-import net.gini.android.internal.payment.utils.extensions.getLayoutInflaterWithGiniPaymentTheme
-import net.gini.android.internal.payment.utils.extensions.hideErrorMessage
-import net.gini.android.internal.payment.utils.extensions.hideKeyboard
-import net.gini.android.internal.payment.utils.extensions.setErrorMessage
-import net.gini.android.internal.payment.utils.extensions.showErrorMessage
-import net.gini.android.internal.payment.utils.setBackgroundTint
-import net.gini.android.internal.payment.utils.setTextIfDifferent
+import net.gini.android.internal.payment.util.amountWatcher
+import net.gini.android.internal.payment.util.extensions.clearErrorMessage
+import net.gini.android.internal.payment.util.extensions.getLayoutInflaterWithGiniPaymentTheme
+import net.gini.android.internal.payment.util.extensions.hideErrorMessage
+import net.gini.android.internal.payment.util.extensions.hideKeyboard
+import net.gini.android.internal.payment.util.extensions.setErrorMessage
+import net.gini.android.internal.payment.util.extensions.showErrorMessage
+import net.gini.android.internal.payment.util.setBackgroundTint
+import net.gini.android.internal.payment.util.setTextIfDifferent
 import org.slf4j.LoggerFactory
 
 interface ReviewViewListener {
