+++ internal-payment-sdk/sdk/src/main/java/net/gini/android/internal/payment/util/AutoClearedValue.kt
@@ -11,7 +11,7 @@ import kotlin.reflect.KProperty
  *
  * Accessing this variable while the fragment's view is destroyed will throw NPE.
  */
-internal class AutoClearedValue<T : Any>(val fragment: Fragment) : ReadWriteProperty<Fragment, T> {
+class AutoClearedValue<T : Any>(val fragment: Fragment) : ReadWriteProperty<Fragment, T> {
     private var _value: T? = null
 
     init {
@@ -42,4 +42,4 @@ internal class AutoClearedValue<T : Any>(val fragment: Fragment) : ReadWriteProp
 /**
  * Creates an [AutoClearedValue] associated with this fragment.
  */
-internal fun <T : Any> Fragment.autoCleared() = AutoClearedValue<T>(this)
+fun <T : Any> Fragment.autoCleared() = AutoClearedValue<T>(this)
