+++ internal-payment-sdk/sdk/src/main/java/net/gini/android/internal/payment/util/BackListener.kt
@@ -1,4 +1,4 @@
-package net.gini.android.internal.payment.utils
+package net.gini.android.internal.payment.util
 
 /**
  * Listener for back events.
