+++ internal-payment-sdk/sdk/src/main/java/net/gini/android/internal/payment/util/FlowBottomSheetsManager.kt
@@ -1,4 +1,4 @@
-package net.gini.android.internal.payment.utils
+package net.gini.android.internal.payment.util
 
 import kotlinx.coroutines.CoroutineScope
 import kotlinx.coroutines.Dispatchers
