+++ internal-payment-sdk/sdk/src/main/java/net/gini/android/internal/payment/util/GiniPaymentManager.kt
@@ -1,4 +1,4 @@
-package net.gini.android.internal.payment.utils
+package net.gini.android.internal.payment.util
 
 import kotlinx.coroutines.flow.MutableStateFlow
 import kotlinx.coroutines.flow.StateFlow
@@ -11,7 +11,7 @@ import net.gini.android.internal.payment.api.model.toPaymentRequest
 import net.gini.android.internal.payment.paymentprovider.PaymentProviderApp
 import net.gini.android.internal.payment.review.ValidationMessage
 import net.gini.android.internal.payment.review.validate
-import net.gini.android.internal.payment.utils.extensions.toBackendFormat
+import net.gini.android.internal.payment.util.extensions.toBackendFormat
 import org.slf4j.LoggerFactory
 
 /**
