+++ internal-payment-sdk/sdk/src/main/java/net/gini/android/internal/payment/util/PaymentFileProvider.kt
@@ -1,4 +1,4 @@
-package net.gini.android.internal.payment.utils
+package net.gini.android.internal.payment.util
 
 import androidx.core.content.FileProvider
