+++ internal-payment-sdk/sdk/src/main/java/net/gini/android/internal/payment/util/extensions/BottomSheetDialogFragment.kt
@@ -1,8 +1,8 @@
-package net.gini.android.internal.payment.utils.extensions
+package net.gini.android.internal.payment.util.extensions
 
 import androidx.activity.OnBackPressedCallback
 import com.google.android.material.bottomsheet.BottomSheetDialog
-import net.gini.android.internal.payment.utils.BackListener
+import net.gini.android.internal.payment.util.BackListener
 
 internal fun BottomSheetDialog.setBackListener(backListener: BackListener) {
     onBackPressedDispatcher.addCallback(object : OnBackPressedCallback(true) {
