+++ internal-payment-sdk/sdk/src/main/java/net/gini/android/internal/payment/util/extensions/Fragment.kt
@@ -1,8 +1,12 @@
 package net.gini.android.internal.payment.util.extensions
 
+import android.app.PendingIntent
+import android.content.Intent
 import android.view.LayoutInflater
+import androidx.core.content.FileProvider
 import androidx.fragment.app.Fragment
 import net.gini.android.internal.payment.GiniInternalPaymentModule
+import java.io.File
 import java.util.Locale
 
 internal fun Fragment.getLayoutInflaterWithGiniPaymentThemeAndLocale(inflater: LayoutInflater, locale: Locale? = null): LayoutInflater {
@@ -15,4 +19,24 @@ internal fun Fragment.getLocaleStringResource(resourceId: Int, giniPaymentModule
     }
 
     return giniPaymentModule?.localizedContext?.getText(resourceId).toString()
+}
+
+fun Fragment.startSharePdfIntent(file: File, pendingIntent: PendingIntent? = null) {
+    val uriForFile = FileProvider.getUriForFile(
+        requireContext(),
+        requireContext().packageName + ".internal.payment.fileprovider",
+        file
+    )
+    val shareIntent = Intent(Intent.ACTION_SEND).apply {
+        type = "application/pdf"
+        flags = Intent.FLAG_GRANT_READ_URI_PERMISSION
+        putExtra(Intent.EXTRA_STREAM, uriForFile)
+    }
+    startActivity(
+        Intent.createChooser(
+            shareIntent,
+            uriForFile.lastPathSegment,
+            pendingIntent?.intentSender
+        )
+    )
 }
