--- internal-payment-sdk/sdk/src/main/java/net/gini/android/internal/payment/utils/AutoClearedValue.kt
@@ -1,45 +1,0 @@
-package net.gini.android.internal.payment.utils
-
-import androidx.fragment.app.Fragment
-import androidx.lifecycle.DefaultLifecycleObserver
-import androidx.lifecycle.LifecycleOwner
-import kotlin.properties.ReadWriteProperty
-import kotlin.reflect.KProperty
-
-/**
- * A lazy property that gets cleaned up when the fragment's view is destroyed.
- *
- * Accessing this variable while the fragment's view is destroyed will throw NPE.
- */
-class AutoClearedValue<T : Any>(val fragment: Fragment) : ReadWriteProperty<Fragment, T> {
-    private var _value: T? = null
-
-    init {
-        fragment.lifecycle.addObserver(object: DefaultLifecycleObserver {
-            override fun onCreate(owner: LifecycleOwner) {
-                fragment.viewLifecycleOwnerLiveData.observe(fragment) { viewLifecycleOwner ->
-                    viewLifecycleOwner?.lifecycle?.addObserver(object: DefaultLifecycleObserver {
-                        override fun onDestroy(owner: LifecycleOwner) {
-                            _value = null
-                        }
-                    })
-                }
-            }
-        })
-    }
-
-    override fun getValue(thisRef: Fragment, property: KProperty<*>): T {
-        return _value ?: throw IllegalStateException(
-            "should never call auto-cleared-value get when it might not be available"
-        )
-    }
-
-    override fun setValue(thisRef: Fragment, property: KProperty<*>, value: T) {
-        _value = value
-    }
-}
-
-/**
- * Creates an [AutoClearedValue] associated with this fragment.
- */
-fun <T : Any> Fragment.autoCleared() = AutoClearedValue<T>(this)
