+++ merchant-sdk/sdk/src/main/java/net/gini/android/merchant/sdk/integratedFlow/PaymentFragment.kt
@@ -32,12 +32,12 @@ import net.gini.android.internal.payment.paymentprovider.PaymentProviderApp
 import net.gini.android.internal.payment.review.ReviewConfiguration
 import net.gini.android.internal.payment.review.reviewBottomSheet.ReviewBottomSheet
 import net.gini.android.internal.payment.review.reviewComponent.ReviewViewListener
-import net.gini.android.internal.payment.utils.PaymentNextStep
-import net.gini.android.internal.payment.utils.autoCleared
-import net.gini.android.internal.payment.utils.extensions.createShareWithPendingIntent
-import net.gini.android.internal.payment.utils.extensions.showInstallAppBottomSheet
-import net.gini.android.internal.payment.utils.extensions.showOpenWithBottomSheet
-import net.gini.android.internal.payment.utils.extensions.startSharePdfIntent
+import net.gini.android.internal.payment.util.PaymentNextStep
+import net.gini.android.internal.payment.util.autoCleared
+import net.gini.android.internal.payment.util.extensions.createShareWithPendingIntent
+import net.gini.android.internal.payment.util.extensions.showInstallAppBottomSheet
+import net.gini.android.internal.payment.util.extensions.showOpenWithBottomSheet
+import net.gini.android.internal.payment.util.extensions.startSharePdfIntent
 import net.gini.android.merchant.sdk.GiniMerchant
 import net.gini.android.merchant.sdk.R
 import net.gini.android.merchant.sdk.databinding.GmsFragmentMerchantBinding
