+++ merchant-sdk/sdk/src/main/java/net/gini/android/merchant/sdk/paymentComponentBottomSheet/PaymentComponentBottomSheet.kt
@@ -12,10 +12,10 @@ import com.google.android.material.bottomsheet.BottomSheetDialog
 import kotlinx.coroutines.launch
 import net.gini.android.internal.payment.paymentComponent.PaymentComponent
 import net.gini.android.internal.payment.paymentComponent.PaymentComponentView
+import net.gini.android.internal.payment.util.BackListener
+import net.gini.android.internal.payment.util.GpsBottomSheetDialogFragment
+import net.gini.android.internal.payment.util.autoCleared
 import net.gini.android.merchant.sdk.databinding.GmsBottomSheetPaymentComponentBinding
-import net.gini.android.internal.payment.utils.BackListener
-import net.gini.android.internal.payment.utils.GpsBottomSheetDialogFragment
-import net.gini.android.internal.payment.utils.autoCleared
 import net.gini.android.merchant.sdk.util.extensions.setBackListener
 import org.jetbrains.annotations.VisibleForTesting
