+++ merchant-sdk/sdk/src/main/java/net/gini/android/merchant/sdk/paymentComponentBottomSheet/PaymentComponentBottomSheetViewModel.kt
@@ -9,7 +9,7 @@ import kotlinx.coroutines.launch
 import net.gini.android.internal.payment.paymentComponent.PaymentComponent
 import net.gini.android.internal.payment.paymentComponent.SelectedPaymentProviderAppState
 import net.gini.android.internal.payment.paymentprovider.PaymentProviderApp
-import net.gini.android.internal.payment.utils.BackListener
+import net.gini.android.internal.payment.util.BackListener
 
 internal class PaymentComponentBottomSheetViewModel private constructor(val paymentComponent: PaymentComponent?, val backListener: BackListener?, val reviewFragmentShown: Boolean) : ViewModel() {
     private val _paymentProviderApp = MutableStateFlow<PaymentProviderApp?>(null)
