+++ merchant-sdk/sdk/src/main/java/net/gini/android/merchant/sdk/util/extensions/BottomSheetDialogFragment.kt
@@ -2,7 +2,7 @@ package net.gini.android.merchant.sdk.util.extensions
 
 import androidx.activity.OnBackPressedCallback
 import com.google.android.material.bottomsheet.BottomSheetDialog
-import net.gini.android.internal.payment.utils.BackListener
+import net.gini.android.internal.payment.util.BackListener
 
 internal fun BottomSheetDialog.setBackListener(backListener: BackListener) {
     onBackPressedDispatcher.addCallback(object : OnBackPressedCallback(true) {
