+++ merchant-sdk/sdk/src/test/java/net/gini/android/merchant/sdk/integratedFlow/PaymentFlowFragmentTest.kt
@@ -15,7 +15,7 @@ import kotlinx.coroutines.flow.MutableStateFlow
 import kotlinx.coroutines.test.runTest
 import net.gini.android.internal.payment.GiniInternalPaymentModule
 import net.gini.android.internal.payment.paymentComponent.PaymentComponent
-import net.gini.android.internal.payment.utils.PaymentNextStep
+import net.gini.android.internal.payment.util.PaymentNextStep
 import net.gini.android.merchant.sdk.GiniMerchant
 import net.gini.android.merchant.sdk.R
 import net.gini.android.merchant.sdk.util.DisplayedScreen
